requirements.txt
streamlit==1.29.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
nltk==3.8.1
pillow==10.1.0
matplotlib==3.8.2
wordcloud==1.9.3
joblib==1.3.2